+++ src/Migration/Migration1759136504AccessToken.php
@@ -1,0 +1,28 @@
+<?php declare(strict_types=1);
+
+namespace RH\Tweakwise\Migration;
+
+use Doctrine\DBAL\Connection;
+use Shopware\Core\Framework\Migration\MigrationStep;
+
+class Migration1759136504AccessToken extends MigrationStep
+{
+    public function getCreationTimestamp(): int
+    {
+        return 1759136504;
+    }
+
+    public function update(Connection $connection): void
+    {
+        $sql = <<<SQL
+ALTER TABLE `s_plugin_rhae_tweakwise_frontend`
+                ADD `accessToken` VARCHAR(255) NULL;
+SQL;
+        $connection->executeStatement($sql);
+    }
+
+    public function updateDestructive(Connection $connection): void
+    {
+        // implement update destructive
+    }
+}