+++ src/Migration/Migration1761558581BackendSyncEnabled.php
@@ -1,0 +1,28 @@
+<?php declare(strict_types=1);
+
+namespace RH\Tweakwise\Migration;
+
+use Doctrine\DBAL\Connection;
+use Shopware\Core\Framework\Migration\MigrationStep;
+
+class Migration1761558581BackendSyncEnabled extends MigrationStep
+{
+    public function getCreationTimestamp(): int
+    {
+        return 1761558581;
+    }
+
+    public function update(Connection $connection): void
+    {
+        $sql = <<<SQL
+ALTER TABLE `s_plugin_rhae_tweakwise_frontend`
+                ADD `backendSyncEnabled` TINYINT(1) NOT NULL DEFAULT '0';
+SQL;
+        $connection->executeStatement($sql);
+    }
+
+    public function updateDestructive(Connection $connection): void
+    {
+        // implement update destructive
+    }
+}