+++ src/Resources/app/administration/src/module/rhae-tweakwise-frontend/page/rhae-tweakwise-frontend-detail/index.js
@@ -34,10 +34,12 @@ Component.register('rhae-tweakwise-frontend-detail', {
             recommendationsAvailable: false,
             processSuccess: false,
             salesChannelDomainIds: [],
+            backendSyncOptions: []
         };
     },
 
     computed: {
+
         integrationTypes() {
             return [
                 {
@@ -125,9 +127,36 @@ Component.register('rhae-tweakwise-frontend-detail', {
     created() {
         this.getItem();
         this.getSalesChannelDomains();
+        this.getBackendSyncOptions();
     },
 
     methods: {
+        async getBackendSyncOptions() {
+            this.isLoading = true;
+            const token = Shopware.Service('loginService').getToken();
+            const headers = { headers: { Authorization: `Bearer ${token}` } };
+            const httpClient = Shopware.Application.getContainer('init').httpClient;
+            const url = `/_action/rhae-tweakwise/sync-options`;
+            const response = await httpClient.get(url, headers);
+
+            let options;
+            if (response.status !== 200) {
+                options = {
+                    'main': [],
+                    'options': [],
+                    'customFields': []
+                }
+            } else {
+                this.backendSyncOptions = response.data;
+            }
+
+            this.isLoading = false;
+        },
+        isChecked(group, option) {
+            return !!(this.item?.backendSyncProperties
+                && this.item.backendSyncProperties[group]
+                && this.item.backendSyncProperties[group][option]);
+        },
         async checkPossibilities() {
             this.validToken = false;
             this.suggestionsAvailable = false
@@ -179,6 +208,18 @@ Component.register('rhae-tweakwise-frontend-detail', {
                 .then((entity) => {
                     this.item = entity;
                     this.checkPossibilities();
+                    // Zorg dat het hoofd-object altijd een object is
+                    if (!this.item.backendSyncProperties || Array.isArray(this.item.backendSyncProperties)) {
+                        this.item.backendSyncProperties = {};
+                    }
+
+                    // Zorg dat de groepen objecten zijn
+                    const groups = ['main', 'properties', 'customFields'];
+                    groups.forEach((g) => {
+                        if (!this.item.backendSyncProperties[g] || Array.isArray(this.item.backendSyncProperties[g])) {
+                            this.item.backendSyncProperties[g] = {};
+                        }
+                    });
                 });
         },
         onChangeValue(value, fieldName, valueChange = true) {
@@ -186,25 +227,50 @@ Component.register('rhae-tweakwise-frontend-detail', {
 
             this.$emit('change-value', fieldName, value);
         },
+        setSelectedProperty(value, fieldName, group, valueChange = true) {
+            if (!this.item.backendSyncProperties || Array.isArray(this.item.backendSyncProperties)) {
+                this.item.backendSyncProperties = {};
+            }
+            if (!this.item.backendSyncProperties[group] || Array.isArray(this.item.backendSyncProperties[group])) {
+                this.item.backendSyncProperties[group] = {};
+            }
+
+            this.item.backendSyncProperties[group][String(fieldName)] = !!value;
+
+            this.$emit('change-value', fieldName, value);
+        },
+        ensureOptionKeysExist() {
+            const groups = ['main', 'properties', 'customFields'];
+            groups.forEach(g => {
+                if (!this.item.backendSyncProperties[g]) this.$set(this.item.backendSyncProperties, g, {});
+                const list = this.backendSyncOptions?.[g] || [];
+                list.forEach(opt => {
+                    if (typeof this.item.backendSyncProperties[g][opt] === 'undefined') {
+                        this.$set(this.item.backendSyncProperties[g], opt, false);
+                    }
+                });
+            });
+        },
         onChangeToggle(value, fieldName) {
             this.onChangeValue(value, fieldName, false);
         },
-        onClickSave() {
+        async onClickSave() {
             this.isLoading = true;
-            this.repository
-                .save(this.item, Context.api)
-                .then(() => {
-                    this.getItem();
-                    this.isLoading = false;
-                    this.processSuccess = true;
-                })
-                .catch((exception) => {
-                    this.isLoading = false;
-                    this.createNotificationError({
-                        title: this.$tc('rhae-tweakwise-frontend.notification.errorTitle'),
-                        message: exception
-                    });
+
+            // this.item.backendSyncProperties = JSON.stringify(this.item.backendSyncProperties);
+
+            try {
+                await this.repository.save(this.item, Context.api);
+                await this.getItem();
+                this.processSuccess = true;
+            } catch (exception) {
+                this.createNotificationError({
+                    title: this.$tc('rhae-tweakwise-frontend.notification.errorTitle'),
+                    message: exception
                 });
+            } finally {
+                this.isLoading = false;
+            }
         },
         setSalesChannelDomains(salesChannelDomains) {
             this.salesChannelDomains = salesChannelDomains;