+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - An option is introduced in the feed settings to respect the Shopware setting to hide products after clearance. When enabled, products and variants will be excluded from the feed after clearance.
+- You can now sync the data of a specific product to Tweakwise from the product editing screen in de admin of Shopware. To be able to use this feature, you need to enable it in the Tweakwise settings in Shopware.
 
 ### Fixed
 - Remove products that are sold out from feed in non-grouped mode if sold out products should be excluded.